048 review remediation 2,3,4,6,7 + fold-ins: reverse-race documented, quickstart filter verified (9/9), BC-teacher changelog line, [rng-sequence] by consequence (owner option 1), strip-witness dormancy prose, is_conscriptable_friend one-body, cycle E recorded

+++ crates/cloudkitty-core/src/behavior/needs_driven.rs
@@ -692,7 +692,7 @@ mod tests {
             world.tick = 20;
             for (id, partner) in [(1, 2), (2, 1)] {
                 let idx = world.kitty_index(id).unwrap();
-                world.kitties[idx].pos = Position::new(5, 4 + id as u32);
+                world.kitties[idx].pos = Position::new(5, 4 + id);
                 world.kitties[idx].needs.add(NeedKind::Play, 60.0);
                 world.kitties[idx].activity = crate::kitty::Activity::Playing {
                     target: Some(crate::action::TargetRef::Kitty { id: partner }),
